App Admin and Permission (#45)
* Keep migration data

* Admin migration

* Make api signatures uniform

* Deduplicate column names

* split case

* Better signature

* stupid

* compromise

* bump types

* Validate project and org

* Validation for task

* User validation

* Notification update

* Fix tests

* move to service

* Admin allow

* Admin sidebar

* mimin boleh baca

* move

+++ backend/src/modules/chatroom/ChatroomService.ts
@@ -1,5 +1,6 @@
-import { NotFoundError } from "../../helpers/error.js";
+import { AuthError, NotFoundError } from "../../helpers/error.js";
 import { ProjectService } from "../project/ProjectService.js";
+import { UserService } from "../user/UserService.js";
 import { ChatRepository } from "./ChatroomRepository.js";
 
 // Kalau chatroomnya berkaitan dengan projek, validasi pakai daftar member projek.
@@ -7,9 +8,11 @@ import { ChatRepository } from "./ChatroomRepository.js";
 export class ChatService {
   repo: ChatRepository;
   project_service: ProjectService;
-  constructor(repo: ChatRepository, project_service: ProjectService) {
+  user_service: UserService;
+  constructor(repo: ChatRepository, project_service: ProjectService, user_service: UserService) {
     this.repo = repo;
     this.project_service = project_service;
+    this.user_service = user_service;
   }
 
   async getMembers(chatroom_id: number) {
@@ -28,9 +31,20 @@ export class ChatService {
     }
   }
 
+  async getAllowedListeners(chatroom_id: number) {
+    const members = await this.getMembers(chatroom_id);
+    const admins = await this.user_service.getUsers({ is_admin: true });
+    return [...members, ...admins.map((x) => x.user_id)];
+  }
+
   async isAllowed(chatroom_id: number, user_id: number) {
     const members = await this.getMembers(chatroom_id);
-    return members.some((member_id) => member_id === user_id);
+    const is_member = members.some((member_id) => member_id === user_id);
+    if (is_member) {
+      return true;
+    }
+    const is_admin = await this.user_service.isAdminUser(user_id);
+    return is_admin;
   }
 
   async getMessages(chatroom_id: number) {
@@ -38,13 +52,28 @@ export class ChatService {
   }
 
   async sendMessage(chatroom_id: number, sender_id: number, message: string) {
+    const val = await this.isAllowed(chatroom_id, sender_id);
+    if (!val) {
+      throw new AuthError("Anda tidak memiliki akses untuk mengirim ke chat ini!");
+    }
+
     return await this.repo.addMessage(chatroom_id, sender_id, message);
   }
 
   async updateMessage(
     message_id: number,
     obj: { chatroom_id?: number; sender_id?: number; message?: string },
+    sender_id: number,
   ) {
+    const old_message = await this.getMessage(message_id);
+    if (!old_message) {
+      throw new NotFoundError("Pesan tidak ditemukan!");
+    }
+
+    if (old_message.user_id !== sender_id) {
+      throw new AuthError("Anda tidak memiliki akses untuk mengubah pesan ini!");
+    }
+
     return await this.repo.updateMessage(message_id, {
       ...obj,
       is_edited: true,
@@ -63,7 +92,12 @@ export class ChatService {
     return await this.repo.getProjectChatrooms(project_id);
   }
 
-  async addProjectChatroom(project_id: number, chatroom_name: string) {
+  async addProjectChatroom(project_id: number, chatroom_name: string, sender_id: number) {
+    const member_role = await this.project_service.getMemberRole(project_id, sender_id);
+    if (member_role !== "Admin" && member_role !== "Dev") {
+      throw new AuthError("Anda tidak memiliki akses untuk melakukan hal ini!");
+    }
+
     return await this.repo.addProjectChatroom(project_id, chatroom_name);
   }
 
@@ -71,15 +105,32 @@ export class ChatService {
     return await this.repo.getUserChatrooms(user_id);
   }
 
-  async addUserChatroom(user_id: number, chatroom_name: string) {
+  async addUserChatroom(user_id: number, chatroom_name: string, sender_id: number) {
+    if (sender_id != user_id) {
+      throw new AuthError("Anda tidak memiliki akses untuk menambahkan chatroom orang lain!");
+    }
+
     return await this.repo.addUserChatroom(user_id, chatroom_name);
   }
 
-  async updateChatroom(chatroom_id: number, name?: string, user_ids?: number[]) {
-    await this.repo.updateChatroom(chatroom_id, name, user_ids);
+  async updateChatroom(
+    chatroom_id: number,
+    opts: { name?: string; user_ids?: number[] },
+    sender_id: number,
+  ) {
+    const val = await this.isAllowed(chatroom_id, sender_id);
+    if (!val) {
+      throw new AuthError("Anda tidak memiliki akses untuk mengirim ke chat ini!");
+    }
+
+    await this.repo.updateChatroom(chatroom_id, opts);
   }
 
-  async deleteChatroom(chatroom_id: number) {
+  async deleteChatroom(chatroom_id: number, sender_id: number) {
+    const val = await this.isAllowed(chatroom_id, sender_id);
+    if (!val) {
+      throw new AuthError("Anda tidak memiliki akses untuk mengirim ke chat ini!");
+    }
     await this.repo.deleteChatroom(chatroom_id);
   }
 }
