+++ backend/src/test/setup-test.ts
@@ -41,6 +41,6 @@ export async function clearDB(app: Application) {
   await app.db.deleteFrom("categories_orgs").execute();
   await app.db.deleteFrom("ms_category_orgs").execute();
   await app.db.deleteFrom("ms_orgs").execute();
-  await app.db.deleteFrom("ms_users").execute();
+  await app.db.deleteFrom("ms_users").where("ms_users.is_admin", "=", false).execute();
   await app.db.deleteFrom("session").execute();
 }
